requirements.txt
streamlit
anthropic 
google-generativeai 
openrouter
python-dotenv
